Ignore frames that have unknown buffer format.
It seems that on HDR once in a while we get
DXGI_FORMAT_R16G16B16A16_FLOAT even if we only should get regular RGB.

This happens when the duplication has an update ONLY on the mouse
position and not on the actual image. It seems that then the system
doesn't bother to make any buffer conversions and we can use the old
image we acquired before.

The lack of conversion messes up the color, as we append -1 to the buffer and it effects
with white flashes once in a while.