Resume grabbing also on successful device calls
The LedDeviceManager emits ioDeviceSuccess for ledDeviceCommandCompleted, but from there it does not trigger a re-open because it isn't actually an I/O failure.
I'd rather not stop emitting the signal there as it has been like this for 11 years, so instead we simply resume grabbing if an error was only temporary.